split constants.py
