+++ src/bequem/nodes/constants/constant_unitary.py
@@ -2,59 +2,8 @@ import numpy as np
 
 from bequem.circuit import Circuit
 from bequem.circuits.generic_unitary import generic_unitary
-from bequem.circuits.state_prep import prepare_state
 from bequem.nodes.node import Node
-from bequem.subspace import Subspace
-
-
-class ConstantVector(Node):
-    """
-    Node representing the given vector
-
-    :ivar vec:
-        The vector represented by this node
-    """
-
-    vec: np.ndarray
-
-    def __init__(self, vec: np.ndarray):
-        """
-        :param vec:
-            The vector represented by this node
-        """
-        self.n_qubits = round(np.log2(vec.shape[0]))
-        assert 2**self.n_qubits == vec.shape[0]
-        self.vec = vec
-
-    def children(self) -> list[Node]:
-        return []
-
-    def parameters(self) -> dict:
-        return {"vec": self.vec}
-
-    def _subspace_in(self) -> Subspace:
-        return Subspace(0, self.n_qubits)
-
-    def _subspace_out(self) -> Subspace:
-        return Subspace(self.n_qubits)
-
-    def _normalization(self) -> float:
-        return np.linalg.norm(self.vec)
-
-    def compute(self, input: np.ndarray) -> np.ndarray:
-        if input.ndim == 1:
-            return self.vec * input[0]
-        else:
-            return (np.array([self.vec]).T @ input.T).T
-
-    def compute_adjoint(self, input: np.ndarray) -> np.ndarray:
-        return (np.array([np.conj(self.vec)]) @ input.T).T
-
-    def _circuit(self) -> Circuit:
-        normalized = self.vec / self.normalization
-        # reversed because prepare_state expects MSB ordering
-        tq_circuit = prepare_state(normalized, list(reversed(range(self.n_qubits))))
-        return Circuit(tq_circuit)
+from bequem.subspace.subspace import Subspace
 
 
 class ConstantUnitary(Node):
